+++ src/main/java/TRaMis8khae/starbucks/StarbucksApplication.java
@@ -2,8 +2,11 @@ package TRaMis8khae.starbucks;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
-@SpringBootApplication
+@EnableJpaAuditing
+@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
 public class StarbucksApplication {
 
 	public static void main(String[] args) {
