--- src/main/java/TRaMis8khae/starbucks/order/application/OrderDeliveryAddressListService.java
@@ -1,6 +1,0 @@
-package TRaMis8khae.starbucks.order.application;
-
-public interface OrderDeliveryAddressListService {
-
-    public void createOrderDeliveryAddressList();
-}
