--- src/main/java/TRaMis8khae/starbucks/order/application/OrderDeliveryAddressListServiceImpl.java
@@ -1,18 +1,0 @@
-package TRaMis8khae.starbucks.order.application;
-
-import TRaMis8khae.starbucks.order.infrastructure.OrderDeliveryAddressListRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class OrderDeliveryAddressListServiceImpl implements OrderDeliveryAddressListService {
-
-    private final OrderDeliveryAddressListRepository orderDeliveryAddressListRepository;
-
-
-    @Override
-    public void createOrderDeliveryAddressList() {
-
-    }
-}
