--- src/main/java/TRaMis8khae/starbucks/order/application/OrderService.java
@@ -1,7 +1,0 @@
-package TRaMis8khae.starbucks.order.application;
-
-import TRaMis8khae.starbucks.order.vo.OrderRequestVo;
-
-public interface OrderService {
-    void addOrder(OrderRequestVo orderRequestVo);
-}
