--- src/main/java/TRaMis8khae/starbucks/order/dto/OrderResponseDto.java
@@ -1,4 +1,0 @@
-package TRaMis8khae.starbucks.order.dto;
-
-public class OrderResponseDto {
-}
