--- src/main/java/TRaMis8khae/starbucks/order/entity/Coupon.java
@@ -1,24 +1,0 @@
-package TRaMis8khae.starbucks.order.entity;
-
-import jakarta.persistence.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-public class Coupon {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @Column(nullable = false, length = 50)
-    private String couponTitle;
-    private String couponCode;
-    @Column(length = 20)
-    private String discountType;
-    private Integer discount;
-    private String couponType;
-    private LocalDateTime startDateTime;
-    private LocalDateTime endDateTime;
-
-    @Column(name = "`condition`", length = 50)  // condition은 예약어이므로 백틱으로 감싸줌
-    private String condition;
-}
