--- src/main/java/TRaMis8khae/starbucks/order/entity/MobileVoucher.java
@@ -1,15 +1,0 @@
-package TRaMis8khae.starbucks.order.entity;
-
-import jakarta.persistence.*;
-
-import java.time.LocalDate;
-
-@Entity
-public class MobileVoucher {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @Column(length = 50)
-    private String VoucherName;
-    private LocalDate endDate;
-}
