--- src/main/java/TRaMis8khae/starbucks/order/infrastructure/OrderDeliveryAddressListRepository.java
@@ -1,7 +1,0 @@
-package TRaMis8khae.starbucks.order.infrastructure;
-
-import TRaMis8khae.starbucks.order.entity.OrderDeliveryAddressList;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderDeliveryAddressListRepository extends JpaRepository<OrderDeliveryAddressList, Long> {
-}
