--- src/main/java/TRaMis8khae/starbucks/order/infrastructure/dynamic/OrderRepositoryCustom.java
@@ -1,4 +1,0 @@
-package TRaMis8khae.starbucks.order.infrastructure.dynamic;
-
-public interface OrderRepositoryCustom {
-}
