--- src/main/java/TRaMis8khae/starbucks/order/infrastructure/dynamic/OrderRepositoryCustomImpl.java
@@ -1,7 +1,0 @@
-package TRaMis8khae.starbucks.order.infrastructure.dynamic;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class OrderRepositoryCustomImpl implements OrderRepositoryCustom {
-}
