--- src/main/java/TRaMis8khae/starbucks/order/vo/OrderRequestVo.java
@@ -1,17 +1,0 @@
-package TRaMis8khae.starbucks.order.vo;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class OrderRequestVo {
-
-    private String memberUUID;
-    private Double deliveryPrice; // 상품주문리스트로 이동?
-    private Double totalPrice; // 상품주문리스트로 이동?
-    private String cardInfo;
-    private String memberName;
-    private String memberPhone;
-
-}
