--- src/main/java/TRaMis8khae/starbucks/order/vo/OrderResponseVo.java
@@ -1,7 +1,0 @@
-package TRaMis8khae.starbucks.order.vo;
-
-import lombok.Builder;
-
-@Builder
-public class OrderResponseVo {
-}
