+++ src/main/java/TRaMis8khae/starbucks/purchase/application/OrderDeliveryAddressService.java
@@ -1,0 +1,6 @@
+package TRaMis8khae.starbucks.purchase.application;
+
+public interface OrderDeliveryAddressService {
+
+    public void addOrderDeliveryAddress();
+}
