+++ src/main/java/TRaMis8khae/starbucks/purchase/application/OrderService.java
@@ -1,0 +1,8 @@
+package TRaMis8khae.starbucks.purchase.application;
+
+import TRaMis8khae.starbucks.purchase.vo.OrderRequestVo;
+
+public interface OrderService {
+
+    void addOrder(OrderRequestVo orderRequestVo);
+}
