+++ src/main/java/TRaMis8khae/starbucks/purchase/dto/OrderResponseDto.java
@@ -1,0 +1,4 @@
+package TRaMis8khae.starbucks.purchase.dto;
+
+public class OrderResponseDto {
+}
