+++ src/main/java/TRaMis8khae/starbucks/purchase/infrastructure/dynamic/OrderRepositoryCustom.java
@@ -1,0 +1,4 @@
+package TRaMis8khae.starbucks.purchase.infrastructure.dynamic;
+
+public interface OrderRepositoryCustom {
+}
