+++ src/main/java/TRaMis8khae/starbucks/purchase/infrastructure/dynamic/OrderRepositoryCustomImpl.java
@@ -1,0 +1,7 @@
+package TRaMis8khae.starbucks.purchase.infrastructure.dynamic;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class OrderRepositoryCustomImpl implements OrderRepositoryCustom {
+}
