+++ src/main/java/TRaMis8khae/starbucks/purchase/vo/OrderResponseVo.java
@@ -1,0 +1,7 @@
+package TRaMis8khae.starbucks.purchase.vo;
+
+import lombok.Builder;
+
+@Builder
+public class OrderResponseVo {
+}
