+++ src/main/java/TRaMis8khae/starbucks/review/entity/ReviewMedia.java
@@ -7,9 +7,12 @@ import jakarta.persistence.Id;
 
 @Entity
 public class ReviewMedia {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
     private String reviewMediaType;
+
     private String reviewMediaPath;
 }
